FileSharing Website/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Werkzeug
